feat(ui-packs): start with ui-packs
add default ui pack config
