+++ sdks/web-sdk/src/lib/configuration/translation.json
