+++ sdks/web-sdk/src/lib/contexts/configuration/store.ts
@@ -1,6 +1,6 @@
 import { writable } from 'svelte/store';
 import { IAppConfiguration } from './types';
-import { configuration as defaultConfiguration } from '../../default-configuration/configuration';
+import { configuration as defaultConfiguration } from '../../configuration/configuration';
 
 export const configuration = writable<IAppConfiguration>(defaultConfiguration);
