+++ sdks/web-sdk/src/lib/pages/CheckSelfie.svelte
@@ -12,7 +12,7 @@
   import { mergeStepConfig } from '../services/merge-service';
   import { injectPrimaryIntoLayoutGradient } from '../services/theme-manager';
   import { selfieUri, currentStepId, appState } from '../contexts/app-state';
-  import { checkSelfieStep, layout } from '../default-configuration/theme';
+  import { checkSelfieStep, layout } from '../ui-packs/default/theme';
   import {
     EActionNames,
     sendButtonClickEvent,
@@ -26,7 +26,7 @@
 
   const style = makeStylesFromConfiguration(
     merge(
-      injectPrimaryIntoLayoutGradient(layout, $configuration.general.colors.primary),
+      injectPrimaryIntoLayoutGradient($uiPack.layout || {}, $uiPack.general.colors.primary),
       $configuration.layout || {},
     ),
     step.style,
