+++ sdks/web-sdk/src/lib/pages/DocumentPhotoBack.svelte
@@ -18,9 +18,9 @@
   import { currentStepId, appState } from '../contexts/app-state';
   import { documents, selectedDocumentInfo } from '../contexts/app-state/stores';
   import { updateDocument } from '../utils/photo-utils';
-  import { documentPhotoBackStep, settings } from '../default-configuration/theme';
+  import { documentPhotoBackStep, settings } from '../ui-packs/default/theme';
   import merge from 'deepmerge';
-  import { layout } from '../default-configuration/theme';
+  import { layout } from '../ui-packs/default/theme';
   import { DocumentType } from '../contexts/app-state';
   import { preloadNextStepByCurrent } from '../services/preload-service';
   import { mergeStepConfig } from '../services/merge-service';
@@ -40,7 +40,7 @@
 
   const style = makeStylesFromConfiguration(
     merge(
-      injectPrimaryIntoLayoutGradient(layout, $configuration.general.colors.primary),
+      injectPrimaryIntoLayoutGradient($uiPack.layout || {}, $uiPack.general.colors.primary),
       $configuration.layout || {},
     ),
     step.style,
