+++ sdks/web-sdk/src/lib/pages/DocumentPhotoBackStart.svelte
@@ -10,7 +10,7 @@
   import { isNativeCamera } from '../contexts/flows';
   import { documents, selectedDocumentInfo } from '../contexts/app-state/stores';
   import merge from 'deepmerge';
-  import { documentPhotoBackStartStep, layout } from '../default-configuration/theme';
+  import { documentPhotoBackStartStep, layout } from '../ui-packs/default/theme';
   import { mergeStepConfig } from '../services/merge-service';
   import { preloadNextStepByCurrent } from '../services/preload-service';
   import { injectPrimaryIntoLayoutGradient } from '../services/theme-manager';
@@ -27,7 +27,7 @@
 
   const style = makeStylesFromConfiguration(
     merge(
-      injectPrimaryIntoLayoutGradient(layout, $configuration.general.colors.primary),
+      injectPrimaryIntoLayoutGradient($uiPack.layout || {}, $uiPack.general.colors.primary),
       $configuration.layout || {},
     ),
     step.style,
