+++ sdks/web-sdk/src/lib/pages/DocumentSelection.svelte
@@ -7,7 +7,7 @@
   import { DocumentOptions } from '../organisms';
   import { goToPrevStep } from '../contexts/navigation';
   import merge from 'deepmerge';
-  import { documentSelectionStep, layout } from '../default-configuration/theme';
+  import { documentSelectionStep, layout } from '../ui-packs/default/theme';
   import { mergeStepConfig } from '../services/merge-service';
   import { preloadNextStepByCurrent } from '../services/preload-service';
   import { injectPrimaryIntoLayoutGradient } from '../services/theme-manager';
@@ -25,7 +25,7 @@
 
   const style = makeStylesFromConfiguration(
     merge(
-      injectPrimaryIntoLayoutGradient(layout, $configuration.general.colors.primary),
+      injectPrimaryIntoLayoutGradient($uiPack.layout || {}, $uiPack.general.colors.primary),
       $configuration.layout || {},
     ),
     step.style,
