+++ sdks/web-sdk/src/lib/pages/Final.svelte
@@ -14,7 +14,7 @@
     sendFlowCompleteEvent,
     EVerificationStatuses,
   } from '../utils/event-service';
-  import { finalStep, layout } from '../default-configuration/theme';
+  import { finalStep, layout } from '../ui-packs/default/theme';
   import { DecisionStatus } from '../contexts/app-state/types';
   import { mergeStepConfig } from '../services/merge-service';
   import { injectPrimaryIntoLayoutGradient } from '../services/theme-manager';
@@ -27,7 +27,7 @@
 
   const style = makeStylesFromConfiguration(
     merge(
-      injectPrimaryIntoLayoutGradient(layout, $configuration.general.colors.primary),
+      injectPrimaryIntoLayoutGradient($uiPack.layout || {}, $uiPack.general.colors.primary),
       $configuration.layout || {},
     ),
     step.style,
