+++ sdks/web-sdk/src/lib/ui-packs/index.ts
@@ -1,0 +1,1 @@
+export { uiPack } from "./store";
