+++ sdks/web-sdk/src/lib/ui-packs/store.ts
@@ -1,0 +1,6 @@
+import { writable } from 'svelte/store';
+import { IAppConfigurationUI } from '../contexts/configuration';
+import { uiTheme } from './default/theme'
+
+export const uiPack = writable<IAppConfigurationUI>(uiTheme);
+
